Added missing bits of code in LiveSync. (Now everthing is there and somehow connected)
git-svn-id: http://linkedgeodata.googlecode.com/svn/trunk@174 00633da9-98c1-349d-95eb-89b68c819386

+++ src/main/java/org/linkedgeodata/dao/nodestore/NodePositionDAO.java
@@ -150,13 +150,17 @@ public class NodePositionDAO
 	
 	
 	
-	public void remove(Iterable<Long> ids)
+	public void remove(Collection<Long> ids)
 		throws SQLException
 	{
+		if(ids.isEmpty())
+			return;
+		
 		//String idStr = StringUtil.implode(",", SQLUtil.quotePostgres(ids));
 		String idStr = StringUtil.implode(",", ids);
 		
 		String sql = "DELETE FROM " + tableName + " WHERE node_id IN (" + idStr + ")"; 
+		logger.trace(sql);
 		SQLUtil.execute(conn, sql, Void.class);
 	}
